Cleaned up tests
This patch cleans up the tests, removing some duplication, and in
general making things easier to follow.
